requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
langgraph>=0.2.0
langsmith>=0.1.0
langchain-core>=0.3.0
transformers>=4.45.0
accelerate>=0.34.0
bitsandbytes>=0.44.0
huggingface-hub>=0.25.0
qwen-vl-utils>=0.0.8
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
pillow>=10.0.0
httpx>=0.27.0
starlette>=0.41.0
reportlab>=4.2.0
fpdf2>=2.7.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
streamlit>=1.38.0
matplotlib>=3.9.0
seaborn>=0.13.0
numpy>=1.26.0